feat: 포인트 타입 수정 (Long -> Int)

+++ src/main/kotlin/com/marketboro/controller/req/PointTransactionReq.kt
@@ -1,5 +1,5 @@
 package com.marketboro.controller.req
 
 data class PointTransactionReq(
-    val points: Long
+    val points: Int
 )
